buildroot/board: Add default files and scripts
This adds and installs an interfaces file. wpa_supplicant.conf
is populated with dummy values but is made present. The
interfaces file will try for about 1 minute to get an IP

Also adds an init script to load WiFi and LCD modules.

Adds startup script for env-collect.sh which loads the ZPU as
well for OpenPIR backlight control

Adds post-build.sh to install all of these to the right places